Add try-each command and deprecate testall - Does exact same as testall, but has a name that accomodates being able   to configure commands
