+++ lib/commands/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   try:         require('./try'),
   'try:testall': require('./testall'),
-  'try:reset':   require('./reset')
+  'try:reset':   require('./reset'),
+  'try:each': require('./try-each')
 };
