+++ lib/commands/testall.js
@@ -1,30 +1,8 @@
 'use strict';
+var TryEachCommand = require('./try-each');
+var extend = require('extend');
 
-module.exports = {
+module.exports = extend({}, TryEachCommand, {
   name: 'try:testall',
-  description: 'Runs `ember test` with each of the dependency scenarios specified in config.' ,
-  works: 'insideProject',
-
-  availableOptions: [
-    { name: 'skip-cleanup',  type: Boolean, default: false },
-    { name: 'config-path', type: String, default: 'config/ember-try.js' }
-  ],
-
-  _getConfig: require('../utils/config'),
-  _TryEachTask: require('../tasks/try-each'),
-
-  run: function(commandOptions, rawArgs) {
-    var config = this._getConfig({
-      project: this.project,
-      configPath: commandOptions.configPath
-    });
-
-    var tryEachTask = new this._TryEachTask({
-      ui: this.ui,
-      project: this.project,
-      config: config
-    });
-
-    return tryEachTask.run(config.scenarios, commandOptions);
-  }
-};
+  description: '(Deprecated, use try:each). Runs `ember test` with each of the dependency scenarios specified in config.'
+});
